api: replace 'long object_size = -1' sentinel with std::optional<uint64_t>
api: replace 'long object_size = -1' sentinel with std::optional<uint64_t>

+++ tests/tests.cc
@@ -461,7 +461,7 @@ class Tests {
       std::string object_name = RandObjectName();
       std::string data = "PutObject()";
       std::stringstream ss(data);
-      minio::s3::PutObjectArgs args(ss, data.length(), 0);
+      minio::s3::PutObjectArgs args(ss, static_cast<long>(data.length()), 0);
       args.bucket = bucket_name_;
       args.object = object_name;
       minio::s3::PutObjectResponse resp = client_.PutObject(args);
@@ -475,7 +475,7 @@ class Tests {
       std::string object_name = RandObjectName();
       size_t size = 67108865;  // (64MiB + 1) bytes
       RandCharStream stream(size);
-      minio::s3::PutObjectArgs args(stream, size, 0);
+      minio::s3::PutObjectArgs args(stream, static_cast<long>(size), 0);
       args.bucket = bucket_name_;
       args.object = object_name;
       minio::s3::PutObjectResponse resp = client_.PutObject(args);
